Revert "move story files outside of pages directory for build optimizations w…"
