+++ frontend/pages/register/register.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { ComponentMeta, ComponentStory } from "@storybook/react"
-import ViewerRegistration from "../pages/register"
+import ViewerRegistration from "."
 
 export default {
   title: "Pages/Registration",
